Spinlock: move definitions from kernel headers to runtime headers
This will allow spinlock code to be compiled for both kernel and
non-kernel usage without having to use `#ifdef KERNEL` directives.
